backend/requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
flask-cors==6.0.2
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gspread==6.2.1
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==26.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
urllib3==2.6.3
Werkzeug==3.1.6
